Add files via upload
Just to update you on what I've tried, basically reducing each pixel into an ITA value and then working out the modal ITA value for each image and trying to use that to train a variety of classifiers to guess skin tone. Think this is a bit of a dead end though. If any bits of the code are unclear, let me know!